+++ .firebase/hosting.c3Jj.cache
@@ -1,0 +1,8 @@
+index.html,1570566015868,368dd3234f32f0c1c86e472f8a068d2a1a3e0b2a98f545a6f11d63d7af9d063b
+main.js,1570565984668,1d0eb0273561039d039e38022c7a06ece32df26692c6c46d0d82e800dec80154
+style.css,1570565984668,ea2b8f783045cf229201267d8f30757fef9c5a536c9a1715813bb7b169410b99
+components/button.js,1570565984668,18038f3b45681ba23000e1b62dedd1926def76b4c4d39601a2802f62c52fc324
+components/input.js,1570565984668,b2e31336b98227730038b07b6fc2920a05ecb4a9ef69d64a4d5552dfee1ff7be
+components/label.js,1570565984668,93f818a5e1f4a457b37cd62ec427ae23b36263b5bf04544b8876ae6d18608b7e
+pages/home.js,1570565984668,e15d59ae861defe1f75b08ef279f5d3a8f2332156c0fbdef92a39232cf078384
+img/mulher.jpg,1570550579521,c09ccd5aa13e10188c7e9004abbe0ea365b2045ff1d3c81667c7ea4b38c2aa34
