+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "rede-social-7c6e9"
+  }
+}
