+++ src/components/button.js
@@ -1,13 +1,15 @@
 function Button(props) {
   const template = `
-    <button class="primary-button" onclick="button.handleClick('${props.id}')" >${props.title}</button>
+    <button class="primary-button" onclick="button.handleClick(event,'${props.id}')" >${props.title}</button>
   `;
   return template;
 }
 
 window.button = {
-  handleClick: (id) => {
-    console.log(`Esse é o meu botão ${id}`);
+  handleClick: (event,id) => {
+    event.preventDefault();
+    
+    // console.log(`Esse é o meu botão ${id}`);
     return `Esse é o meu botão ${id}`;
   },
 };
