+++ src/index.html
@@ -6,10 +6,11 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Rede Social da Segurança</title>
   <link rel="stylesheet" href="style.css">
-oioi
+  <script src="https://cdn.firebase.com/libs/firebaseui/3.5.2/firebaseui.js"></script>
+<link type="text/css" rel="stylesheet" href="https://cdn.firebase.com/libs/firebaseui/3.5.2/firebaseui.css" />
 </head>
 <body>
-<main></main>
+<main id='container'></main>
   <script type="module" src="./main.js"></script>
 </body>
 <script src="/__/firebase/6.4.1/firebase-app.js"></script>
