+++ src/pages/home.js
@@ -1,43 +1,30 @@
 import Button from '../components/button.js';
 import Input from '../components/input.js';
-import Label from '../components/label.js';
+// import Label from '../components/label.js';
 
-function Home(){
-  const template = `
-      <h2> Crie a sua conta </h2>
-<form> 
-     ${Input(
-       { placeholder: 'Nome completo',
-       type: 'text'}
-     )}
-     ${Input(
-       { placeholder: 'E-mail',
-       type: 'email'}
-     )}
-     ${Input(
-       { placeholder: 'Senha',
-       type: 'password'}
-     )}
-     ${Input(
-       { placeholder: 'Data de Nascimento',
-       type: 'number'}
-     )}
-     ${Label(
-       { message: 'Gênero'}
-     )}
-     ${Input(
-       { type: 'radio'}
-     )}
-
-     ${Button(
-       { id: 'button',
-       title: 'Cadastrar' }
-     )}
-</form>
-     `;
-    return template;
+function Login() {
+  const template= `
+  <h2> Faça login na sua conta </h2>
+  <form>
+  ${Input({placeholder: 'Digite seu e-mail', type: 'email'})}
+  ${Input({placeholder: 'Digite sua senha', type: 'password'})}
+  ${Button({id: 'button', title: 'Entrar'})}
+  </form>
+  `;
+  return template;
 }
 
-export default Home;
+export default Login;
+
+firebase.auth().createUserWithEmailAndPassword(email, password).catch(function(error) {
+  // Handle Errors here.
+  var errorCode = error.code;
+  var errorMessage = error.message;
+  // ...
+});email-password.html
+
+
+
+
 
 //${Button({ id: '🎉', title: 'Botão 🎉' })}
